This is just one of my ideas of chopping down trees and getting all the blocks that are a part of that tree.
This is just a working proof of concept, the data is not being saved on
game exit, but if the tree is not found in the tree map, it gets chopped
down with the usual method. I have no idea how good, or bad, this can
get with memory or performace, still needs to be tested, maybe it could
be changed to only track the loaded chunks, that way the map will never
get too big.